Mediate 프로젝트 학교//학년 검색 기능 및 UI 구현 완료 학과 정보 받아오기 기능 구현중(검색하여 얻는 값이 정상적이지 않아 새로운 방법을 찾아야함)
