+++ app/src/main/res/layout/activity_select_info.xml
@@ -19,19 +19,19 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/textView2"
+        android:id="@+id/tv_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
-        android:text="@string/search_school"
+        android:text="@string/select_info_school"
         android:textColor="@color/black"
         android:textSize="16sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <AutoCompleteTextView
-        android:id="@+id/autoCompleteTextView_school"
+    <TextView
+        android:id="@+id/tv_school"
         android:background="@drawable/login_edit_change"
         android:layout_width="match_parent"
         android:layout_height="56dp"
@@ -41,20 +41,58 @@
         android:layout_marginRight="16dp"
         android:layout_marginLeft="16dp"
         android:textSize="14sp"
-        android:textColorHint="@color/text_grey"
+        android:textColor="@color/text_grey"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView2" />
+        app:layout_constraintTop_toBottomOf="@+id/tv_title" />
 
     <ImageView
         android:layout_width="24dp"
         android:layout_height="24dp"
         android:src="@drawable/research"
         android:layout_marginRight="16dp"
-        app:layout_constraintBottom_toBottomOf="@+id/autoCompleteTextView_school"
-        app:layout_constraintEnd_toEndOf="@+id/autoCompleteTextView_school"
-        app:layout_constraintTop_toTopOf="@+id/autoCompleteTextView_school"
+        app:layout_constraintBottom_toBottomOf="@+id/tv_school"
+        app:layout_constraintEnd_toEndOf="@+id/tv_school"
+        app:layout_constraintTop_toTopOf="@+id/tv_school"
         android:layout_marginEnd="16dp" />
+
+    <AutoCompleteTextView
+        android:id="@+id/autoCompleteTextView_major"
+        android:background="@drawable/login_edit_change"
+        android:layout_width="match_parent"
+        android:layout_height="56dp"
+        android:completionThreshold="1"
+        android:padding="16dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginRight="16dp"
+        android:layout_marginLeft="16dp"
+        android:textSize="14sp"
+        android:textColorHint="@color/text_grey"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/tv_school" />
+
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@drawable/spinner_background"
+        android:layout_marginTop="8dp"
+        android:layout_marginRight="16dp"
+        android:layout_marginLeft="16dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/autoCompleteTextView_major">
+        <Spinner
+            android:id="@+id/spinner_grade"
+            android:layout_width="match_parent"
+            android:layout_height="56dp"
+            android:spinnerMode="dropdown"
+            android:textSize="14sp"
+            android:textColor="@color/text_grey"/>
+    </RelativeLayout>
+
+
+
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/bt_next_inactive"
